reorganized auxiliary scripts to preconversion and postconversion packages
